backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
fastf1==3.4.0
redis==5.0.8
httpx==0.27.2
numpy==2.1.1
scipy==1.14.1
pandas==2.2.3
python-dotenv==1.0.1
websockets==13.1
pydantic==2.9.2
pydantic-settings==2.5.2
